separate utility class
