+++ plugins/com.google.cloud.tools.eclipse.appengine.newproject.maven/src/com/google/cloud/tools/eclipse/appengine/newproject/maven/CreateMavenBasedAppEngineStandardProject.java
@@ -18,7 +18,7 @@ import org.eclipse.ui.actions.WorkspaceModifyOperation;
 import org.eclipse.wst.common.project.facet.core.IFacetedProject;
 import org.eclipse.wst.common.project.facet.core.ProjectFacetsManager;
 
-import com.google.cloud.tools.eclipse.appengine.newproject.CreateAppEngineStandardWtpProject;
+import com.google.cloud.tools.eclipse.appengine.newproject.AppEngineFacet;
 
 import java.lang.reflect.InvocationTargetException;
 import java.text.MessageFormat;
@@ -75,8 +75,8 @@ public class CreateMavenBasedAppEngineStandardProject extends WorkspaceModifyOpe
     for (IProject project : archetypeProjects) {
       IFacetedProject facetedProject = ProjectFacetsManager.create(
           project, true, loopMonitor.newChild(1));
-      CreateAppEngineStandardWtpProject.installAppEngineFacet(facetedProject, loopMonitor.newChild(1));  
-      CreateAppEngineStandardWtpProject.installAppEngineRuntime(facetedProject, loopMonitor.newChild(1));  
+      AppEngineFacet.installAppEngineFacet(facetedProject, loopMonitor.newChild(1));  
+      AppEngineFacet.installAppEngineRuntime(facetedProject, loopMonitor.newChild(1));  
     }
     
     /*
