+++ plugins/com.google.cloud.tools.eclipse.appengine.newproject/src/com/google/cloud/tools/eclipse/appengine/newproject/CreateAppEngineStandardWtpProject.java
@@ -18,38 +18,23 @@ import org.eclipse.jst.j2ee.project.facet.IJ2EEModuleFacetInstallDataModelProper
 import org.eclipse.jst.j2ee.web.project.facet.IWebFacetInstallDataModelProperties;
 import org.eclipse.jst.j2ee.web.project.facet.WebFacetInstallDataModelProvider;
 import org.eclipse.jst.j2ee.web.project.facet.WebFacetUtils;
-import org.eclipse.jst.server.core.FacetUtil;
 import org.eclipse.ui.actions.WorkspaceModifyOperation;
 import org.eclipse.ui.ide.undo.CreateProjectOperation;
 import org.eclipse.wst.common.frameworks.datamodel.DataModelFactory;
 import org.eclipse.wst.common.frameworks.datamodel.IDataModel;
 import org.eclipse.wst.common.project.facet.core.IFacetedProject;
-import org.eclipse.wst.common.project.facet.core.IFacetedProjectWorkingCopy;
-import org.eclipse.wst.common.project.facet.core.IProjectFacet;
-import org.eclipse.wst.common.project.facet.core.IProjectFacetVersion;
 import org.eclipse.wst.common.project.facet.core.ProjectFacetsManager;
-import org.eclipse.wst.common.project.facet.core.runtime.IRuntime;
-import org.eclipse.wst.common.project.facet.core.runtime.RuntimeManager;
-import org.eclipse.wst.server.core.IRuntimeType;
-import org.eclipse.wst.server.core.IRuntimeWorkingCopy;
-import org.eclipse.wst.server.core.ServerCore;
 
 import java.lang.reflect.InvocationTargetException;
 import java.net.URI;
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 /**
 * Utility to make a new Eclipse project with the App Engine Standard facets in the workspace.  
 */
 public class CreateAppEngineStandardWtpProject extends WorkspaceModifyOperation {
   
-  private static final String DEFAULT_RUNTIME_ID = "com.google.cloud.tools.eclipse.runtime.custom";
-  private static final String DEFAULT_RUNTIME_NAME = "App Engine (Custom)";
-  public static final String APPENGINE_FACET_ID = "com.google.cloud.tools.eclipse.appengine.facet";
-
   private final AppEngineStandardProjectConfig config;
   private final IAdaptable uiInfoAdapter;
 
@@ -88,22 +73,13 @@ public class CreateAppEngineStandardWtpProject extends WorkspaceModifyOperation 
       installWebFacet(facetedProject, progress.newChild(10));
       
       // must happen after other two facets because the appengine facet requires them
-      installAppEngineFacet(facetedProject, progress.newChild(10));
-      installAppEngineRuntime(facetedProject, progress.newChild(20));
+      AppEngineFacet.installAppEngineFacet(facetedProject, progress.newChild(10));
+      AppEngineFacet.installAppEngineRuntime(facetedProject, progress.newChild(20));
     } catch (ExecutionException ex) {
       throw new InvocationTargetException(ex, ex.getMessage());
     } finally {
       progress.done();
     }
-  }
-
-  public static void installAppEngineFacet(IFacetedProject facetedProject, IProgressMonitor monitor)
-      throws CoreException {
-    IFacetedProjectWorkingCopy workingCopy = facetedProject.createWorkingCopy();
-    IProjectFacet appEngineFacet = ProjectFacetsManager.getProjectFacet(APPENGINE_FACET_ID);
-    IProjectFacetVersion appEngineFacetVersion = appEngineFacet.getVersion("1");
-    workingCopy.addProjectFacet(appEngineFacetVersion);
-    workingCopy.commitChanges(monitor);
   }
 
   static void installJavaFacet(IFacetedProject facetedProject, IProgressMonitor monitor) 
@@ -124,36 +100,5 @@ public class CreateAppEngineStandardWtpProject extends WorkspaceModifyOperation 
     webModel.setBooleanProperty(IWebFacetInstallDataModelProperties.INSTALL_WEB_LIBRARY, false);
     webModel.setStringProperty(IWebFacetInstallDataModelProperties.CONFIG_FOLDER, "src/main/webapp");
     facetedProject.installProjectFacet(WebFacetUtils.WEB_25, webModel, monitor);
-  }
-
-  public static void installAppEngineRuntime(IFacetedProject project, IProgressMonitor monitor)
-      throws CoreException {
-    Set<IProjectFacetVersion> facets = new HashSet<>();
-    facets.add(WebFacetUtils.WEB_25);
-    Set<IRuntime> runtimes = RuntimeManager.getRuntimes(facets);
-    project.setTargetedRuntimes(runtimes, monitor);
-    
-    if (RuntimeManager.isRuntimeDefined(DEFAULT_RUNTIME_NAME)) {
-      IRuntime appEngineRuntime = RuntimeManager.getRuntime(DEFAULT_RUNTIME_NAME);
-      project.setPrimaryRuntime(appEngineRuntime, monitor);
-    } else { // Create a new App Engine runtime
-      IRuntimeType appEngineRuntimeType =
-          ServerCore.findRuntimeType(DEFAULT_RUNTIME_ID);
-      if (appEngineRuntimeType == null) {
-        throw new NullPointerException("Could not find " + DEFAULT_RUNTIME_NAME + " runtime type");
-      }
-
-      IRuntimeWorkingCopy appEngineRuntimeWorkingCopy 
-          = appEngineRuntimeType.createRuntime(null, monitor);
-      org.eclipse.wst.server.core.IRuntime appEngineServerRuntime 
-          = appEngineRuntimeWorkingCopy.save(true, monitor);
-      IRuntime appEngineFacetRuntime = FacetUtil.getRuntime(appEngineServerRuntime);
-      if (appEngineFacetRuntime == null) {
-        throw new NullPointerException("Could not locate App Engine facet runtime");
-      }
-  
-      project.addTargetedRuntime(appEngineFacetRuntime, monitor);
-      project.setPrimaryRuntime(appEngineFacetRuntime, monitor);
-    }
   }
 }
